Split server library into modules
Add the modules `server::handler` and `server::ws` for handlers and
websocket related stuff to avoid things getting messy with the imports.
