+++ src/bin/server.rs
@@ -1,6 +1,6 @@
 use warp::Filter;
 
-use opcode::server::opcode_handler;
+use opcode::server::handler::opcode_handler;
 
 #[tokio::main]
 async fn main() {
