+++ src/server/ws.rs
@@ -1,0 +1,5 @@
+use warp::ws::WebSocket;
+
+pub async fn client_connection(ws: WebSocket) {
+    println!("Established client connection: {:?}", ws);
+}
